requirements.txt
pandas
numpy
matplotlib
pymatgen
monty
python-ternary
